src/requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.0
asynctest==0.13.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
djangorestframework==3.12.4
frozenlist==1.2.0
future==0.18.2
idna==3.3
line-bot-sdk==2.0.1
multidict==5.2.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uWSGI==2.0.19.1
yapf==0.31.0
yarl==1.7.2
